Move base model objects and interfaces to a better-named module
"utils" means "bag of crap" not "core objects and interfaces for
managing model parsing"

Since these are core objects and interfaces for managing model parsing,
let's make it the base model objects module.
