+++ metricflow/model/objects/base.py
