+++ metricflow/model/objects/user_configured_model.py
@@ -3,7 +3,7 @@ from typing import List
 from metricflow.model.objects.data_source import DataSource
 from metricflow.model.objects.materialization import Materialization
 from metricflow.model.objects.metric import Metric
-from metricflow.model.objects.utils import HashableBaseModel
+from metricflow.model.objects.base import HashableBaseModel
 
 
 class UserConfiguredModel(HashableBaseModel):
